// packages/vite/types.ts
interface ClientBuildOptions {
  srcDir?: string;
  entry?: string;
  nodesSubdir?: string;
  nodeFilePattern?: RegExp;
  name?: string;
  format?: "es" | "iife" | "umd";
  base?: string;
  licensePath?: string;
  locales?: LocalesOptions;
  staticDirs?: {
    icons?: string;
    public?: string;
  };
  external?: string[];
  globals?: Record<string, string>;
  manualChunks?: (id: string) => string | undefined;
  plugins: Plugin[];
}

interface CopyTarget {
  src: string;
  dest: string;
}

interface LocalesOptions {
  docsDir?: string;
  labelsDir?: string;
  languages?: string[];
}

interface LoggerOptions {
  name: string;
  prefix?: string;
}

interface NodeRedLauncherOptions {
  runtime?: {
    port?: number;
    settingsFilepath?: string;
    version?: string;
  };
  restartDelay?: number;
  args?: string[];
}
// ...
interface ServerBuildOptions {
  srcDir?: string;
  entry?: string;
  bundled?: string[];
  types?: boolean;
  nodeTarget?: string;
  plugins: Plugin[];
}